refactor css id
