--- src/CommentBoard.css
@@ -1,29 +1,0 @@
-#scroll {
-  /* font-family: 'Bungee';
-  color: white;
-  font-size: 30px;; */
-  width: 1000px;
-  display: flex;
-  flex-direction: column;
-  overflow-y: auto;
-  border: 1px solid white;
-  height: 200px;
-  flex: 0 0 auto;
-  /* justify-content: flex-end; */
-}
-#chatInput {
-  display: flex;
-  flex-direction: row;
-  position: sticky;
-  bottom: 0px;
-  margin-top: 0px;
-}
-#outlined-secondary{
-  width: 870px;
-  background-color: rgb(88, 88, 88);
-}
-#button {
-  width: 130px;
-  background-color: blue;
-  color: white;
-}
