+++ src/Comments/Comment.js
@@ -6,12 +6,12 @@ import './Comment.css';
 
 function Comment(props) {
   return (
-    <div id='comment'>
+    <div id='commentGroup'>
       <div id='meta'>
         <div><strong>{props.author}</strong></div>
         <div><strong>{props.time}</strong></div>
       </div>
-      <div id='content'>{props.text}</div>
+      <div id='commentText'>{props.text}</div>
     </div>
   );
 }
