--- src/Discover.css
@@ -1,95 +1,0 @@
-
-/* .container{
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  width: 100%;
-  background-color: grey;
-  justify-content: flex-start;
-}
-nav {
-  display: flex;
-  min-width: 240px;
-  border: 2px solid red;
-}
-main {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-content: center;
-  width: 72%;
-}
-
-.row {
-  display: flex;
-  flex-direction: row;
-  border: 2px solid red;
-  border-radius: 8px;
-  min-height: 280px;
-  max-height: 280px;
-  width: 100%;
-  justify-content: space-between;
-  margin-bottom: 2em;
-} */
-
-nav {
-  grid-area: nav;
-  border: 2px solid red;
-  background-color: rgb(26, 25, 25);
-  position: fixed;
-  width: 240px;
-  height: 100%;
-  font-family: 'Bungee Hairline';
-  color: white;
-  font-size: 28px;
-  font-weight: 500;
-}
-
-row1 {
-  grid-area: row1;
-  display: flex;
-  justify-content: space-between;
-  /* border: 2px solid green; */
-}
-row2 {
-  grid-area: row2;
-  display: flex;
-}
-row3 {
-  grid-area: row3;
-  display: flex;
-}
-row4 {
-  grid-area: row4;
-  display: flex;
-}
-row5 {
-  grid-area: row5;
-}
-
-div {
-  border-bottom: 1px solid grey;
-}
-
-
-
-.container {
-  min-width: 100%;
-  background-color: rgb(37, 36, 36);
-  border: 2px solid blue;
-  display: grid;
-  /* grid-column-gap: 2em; */
-  grid-template-columns: 240px 1fr;
-  grid-template-rows: auto;
-  grid-template-areas: 
-  "nav row1"
-  "nav row2"
-  "nav row3"
-  "nav row4"
-  "nav row5";
-  grid-row-gap: 10px;
-}
-
-
-
-
