+++ src/GameCards/GameCard.css
@@ -55,7 +55,7 @@
 
 @font-face {
   font-family: 'Star Trek';
-  src:  url('./img/star.ttf') format('truetype'); /* Chrome 4+, Firefox 3.5, Opera 10+, Safari 3—5 */
+  src:  url('../img/star.ttf') format('truetype'); /* Chrome 4+, Firefox 3.5, Opera 10+, Safari 3—5 */
 }
 #overview {
   font-family: 'Star Trek';
